+++ asusrouter/tools/security.py
@@ -27,18 +27,6 @@ class ARSecurityLevel(FromIntMixin, IntEnum):
     UNSAFE = 9
 
     @classmethod
-    def at_least_strict(cls, level: ARSecurityLevel) -> bool:
-        """Check if the security level is at least strict."""
-
-        return level.value >= cls.STRICT.value
-
-    @classmethod
-    def at_least_default(cls, level: ARSecurityLevel) -> bool:
-        """Check if the security level is at least default."""
-
-        return level.value >= cls.DEFAULT.value
-
-    @classmethod
     def at_least_sanitized(cls, level: ARSecurityLevel) -> bool:
         """Check if the security level is at least sanitized."""